feat: add ReceiptsListPage and ReceiptDetailPage components with functionality for managing receipts
- Implemented ReceiptsListPage for displaying and filtering receipts with search functionality.
- Added ReceiptDetailPage for uploading and managing receipt details, including alias and item photos.
- Created tests for both pages to ensure proper functionality and user interactions.

+++ Backend/src/ProzoroBanka.Application/Admin/DTOs/AdminDtos.cs
@@ -22,7 +22,7 @@ public record AdminOrganizationDto(
 	string OwnerEmail,
 	int MemberCount,
 	int CampaignCount,
-	decimal TotalRaised,
+	long TotalRaised,
 	DateTime CreatedAt,
 	OrganizationPlanType PlanType);
 
@@ -43,8 +43,8 @@ public record AdminCampaignDto(
 	string Title,
 	string? Description,
 	string? CoverImageUrl,
-	decimal GoalAmount,
-	decimal CurrentAmount,
+	long GoalAmount,
+	long CurrentAmount,
 	string? SendUrl,
 	CampaignStatus Status,
 	DateTime? StartDate,
