+++ Backend/src/ProzoroBanka.Application/Campaigns/Commands/CreateCampaign/CreateCampaignCommand.cs
@@ -12,7 +12,7 @@ public record CreateCampaignCommand(
 	Guid OrganizationId,
 	string Title,
 	string? Description,
-	decimal GoalAmount,
+	long GoalAmount,
 	DateTime? Deadline,
 	string? SendUrl) : IRequest<ServiceResponse<CampaignDto>>, ICacheInvalidatingCommand
 {
